fix aggregate
