+++ wren-modeling-rs/core/src/logical_plan/analyze/mod.rs
@@ -1,5 +1,6 @@
+pub mod model_anlayze;
+pub mod model_generation;
 pub mod plan;
 mod relation_chain;
-pub mod rule;
 
 pub use relation_chain::RelationChain;
