+++ wren-modeling-rs/core/src/mdl/context.rs
@@ -2,7 +2,7 @@ use std::any::Any;
 use std::ops::Deref;
 use std::sync::Arc;
 
-use crate::logical_plan::analyze::rule::{ModelAnalyzeRule, ModelGenerationRule};
+use crate::logical_plan::analyze::model_generation::ModelGenerationRule;
 use crate::logical_plan::utils::create_schema;
 use crate::mdl::manifest::Model;
 use crate::mdl::{AnalyzedWrenMDL, WrenMDL};
@@ -19,6 +19,7 @@ use datafusion::physical_plan::ExecutionPlan;
 use datafusion::prelude::SessionContext;
 use datafusion::sql::TableReference;
 use parking_lot::RwLock;
+use crate::logical_plan::analyze::model_anlayze::ModelAnalyzeRule;
 
 /// Apply Wren Rules to the context for sql generation.
 /// TODO: There're some issue for unparsing the datafusion optimized plans.
