+++ wren-modeling-rs/core/src/mdl/dataset.rs
@@ -4,8 +4,8 @@ use crate::mdl::utils::quoted;
 use crate::mdl::{RegisterTables, SessionStateRef};
 use datafusion::arrow::datatypes::DataType::Utf8;
 use datafusion::arrow::datatypes::Field;
-use datafusion::common::{not_impl_err, DFSchema};
 use datafusion::common::Result;
+use datafusion::common::{not_impl_err, DFSchema};
 use datafusion::logical_expr::sqlparser::ast::Expr::CompoundIdentifier;
 use datafusion::sql::sqlparser::ast::Expr::Identifier;
 use datafusion::sql::sqlparser::ast::{visit_expressions, Expr, Ident};
