+++ wren-modeling-rs/wren-example/examples/plan-sql.rs
@@ -1,7 +1,7 @@
 use datafusion::prelude::SessionContext;
 use std::sync::Arc;
 use wren_core::mdl::builder::{
-    ColumnBuilder, ManifestBuilder, ModelBuilder, RelationshipBuilder, ViewBuilder,
+    ColumnBuilder, ManifestBuilder, ModelBuilder, RelationshipBuilder,
 };
 use wren_core::mdl::manifest::{JoinType, Manifest};
 use wren_core::mdl::{transform_sql_with_ctx, AnalyzedWrenMDL};
@@ -11,17 +11,9 @@ async fn main() -> datafusion::common::Result<()> {
     let manifest = init_manifest();
     let analyzed_mdl = Arc::new(AnalyzedWrenMDL::analyze(manifest)?);
 
-    let sql = r#"
-        select count(*) from wrenai.public."Order_items";
-    "#;
+    let sql = "select customer_state from wrenai.public.orders_model";
     println!("Original SQL: \n{}", sql);
-    let sql = match transform_sql_with_ctx(&SessionContext::new(), analyzed_mdl, sql).await {
-        Ok(sql) => sql,
-        Err(e) => {
-            eprintln!("Error: {}", e);
-            return Ok(());
-        }
-    };
+    let sql = transform_sql_with_ctx(&SessionContext::new(), analyzed_mdl, sql).await?;
     println!("Wren engine generated SQL: \n{}", sql);
     Ok(())
 }
@@ -29,119 +21,75 @@ async fn main() -> datafusion::common::Result<()> {
 fn init_manifest() -> Manifest {
     ManifestBuilder::new()
         .model(
-            ModelBuilder::new("Customers")
+            ModelBuilder::new("customers_model")
                 .table_reference("datafusion.public.customers")
-                .column(ColumnBuilder::new("City", "varchar").expression("city").build())
-                .column(ColumnBuilder::new("Id", "varchar").expression("id").build())
-                .column(ColumnBuilder::new("State", "varchar").expression("state").build())
-                .column(
-                    ColumnBuilder::new_calculated("City_state", "varchar")
-                        .expression(r#""City" || ' ' || "State""#)
-                        .build(),
-                )
-                .primary_key("Id")
+                .column(ColumnBuilder::new("city", "varchar").build())
+                .column(ColumnBuilder::new("id", "varchar").build())
+                .column(ColumnBuilder::new("state", "varchar").build())
+                .primary_key("id")
                 .build(),
         )
         .model(
-            ModelBuilder::new("Order_items")
+            ModelBuilder::new("order_items_model")
                 .table_reference("datafusion.public.order_items")
-                .column(ColumnBuilder::new("Freight_value", "double").expression("freight_value").build())
-                .column(ColumnBuilder::new("Id", "bigint").expression("id").build())
-                .column(ColumnBuilder::new("Item_number", "bigint").expression("item_number").build())
-                .column(ColumnBuilder::new("Order_id", "varchar").expression("order_id").build())
-                .column(ColumnBuilder::new("Price", "double").expression("price").build())
-                .column(ColumnBuilder::new("Product_id", "varchar").expression("product_id").build())
-                .column(ColumnBuilder::new("Shipping_limit_date", "varchar").expression("shipping_limit_date").build())
+                .column(ColumnBuilder::new("freight_value", "double").build())
+                .column(ColumnBuilder::new("id", "bigint").build())
+                .column(ColumnBuilder::new("item_number", "bigint").build())
+                .column(ColumnBuilder::new("order_id", "varchar").build())
+                .column(ColumnBuilder::new("price", "double").build())
+                .column(ColumnBuilder::new("product_id", "varchar").build())
+                .column(ColumnBuilder::new("shipping_limit_date", "varchar").build())
                 .column(
-                    ColumnBuilder::new_relationship(
-                        "Orders",
-                        "Orders",
-                        "Orders_order_items",
-                    )
+                    ColumnBuilder::new("orders_model", "orders_model")
+                        .relationship("orders_order_items")
                         .build(),
                 )
                 .column(
-                    ColumnBuilder::new_calculated("Customer_state", "varchar")
-                        .expression(r#""Orders"."Customers"."State""#)
+                    ColumnBuilder::new("customer_state", "varchar")
+                        .calculated(true)
+                        .expression("orders_model.customers_model.state")
                         .build(),
                 )
-                .primary_key("Id")
+                .primary_key("id")
                 .build(),
         )
         .model(
-            ModelBuilder::new("Orders")
+            ModelBuilder::new("orders_model")
                 .table_reference("datafusion.public.orders")
-                .column(ColumnBuilder::new("Approved_timestamp", "varchar").expression("approved_timestamp").build())
-                .column(ColumnBuilder::new("Customer_id", "varchar").expression("customer_id").build())
-                .column(ColumnBuilder::new("Delivered_carrier_date", "varchar").expression("delivered_carrier_date").build())
-                .column(ColumnBuilder::new("Estimated_delivery_date", "varchar").expression("estimated_delivery_date").build())
-                .column(ColumnBuilder::new("Order_id", "varchar").expression("order_id").build())
-                .column(ColumnBuilder::new("Purchase_timestamp", "varchar").expression("purchase_timestamp").build())
+                .column(ColumnBuilder::new("approved_timestamp", "varchar").build())
+                .column(ColumnBuilder::new("customer_id", "varchar").build())
+                .column(ColumnBuilder::new("delivered_carrier_date", "varchar").build())
+                .column(ColumnBuilder::new("estimated_delivery_date", "varchar").build())
+                .column(ColumnBuilder::new("order_id", "varchar").build())
+                .column(ColumnBuilder::new("purchase_timestamp", "varchar").build())
                 .column(
-                    ColumnBuilder::new_relationship(
-                        "Customers",
-                        "Customers",
-                        "Orders_customer",
-                    )
+                    ColumnBuilder::new("customers_model", "customers_model")
+                        .relationship("orders_customer")
                         .build(),
                 )
                 .column(
-                    ColumnBuilder::new_calculated("Customer_state", "varchar")
-                        .expression(r#""Customers"."State""#)
+                    ColumnBuilder::new("customer_state", "varchar")
+                        .calculated(true)
+                        .expression("customers_model.state")
                         .build(),
                 )
-                // TODO: fix calcaultion with non-relationship column
-                // .column(
-                //     ColumnBuilder::new_calculated("Customer_state_order_id", "varchar")
-                //         .expression(r#""Customers"."State" || ' ' || "Order_id""#)
-                //         .build(),
-                // )
-                .column(
-                    ColumnBuilder::new_relationship(
-                        "Order_items",
-                        "Order_items",
-                        "Orders_order_items",
-                    )
-                        .build(),
-                )
-                .column(
-                    ColumnBuilder::new_calculated("Totalprice", "double")
-                        .expression(r#"sum("Order_items"."Price")"#)
-                        .build(),
-                )
-                .column(
-                    ColumnBuilder::new_calculated("Customer_city", "varchar")
-                        .expression(r#""Customers"."City""#)
-                        .build(),
-                )
-                .primary_key("Order_id")
                 .build(),
         )
         .relationship(
-            RelationshipBuilder::new("Orders_customer")
-                .model("Orders")
-                .model("Customers")
+            RelationshipBuilder::new("orders_customer")
+                .model("orders_model")
+                .model("customers_model")
                 .join_type(JoinType::ManyToOne)
-                .condition(r#""Orders"."Customer_id" = "Customers"."Id""#)
+                .condition("orders_model.customer_id = customers_model.id")
                 .build(),
         )
         .relationship(
-            RelationshipBuilder::new("Orders_order_items")
-                .model("Orders")
-                .model("Order_items")
+            RelationshipBuilder::new("orders_order_items")
+                .model("orders_model")
+                .model("order_items_model")
                 .join_type(JoinType::ManyToOne)
-                .condition(r#""Orders"."Order_id" = "Order_items"."Order_id""#)
+                .condition("orders_model.order_id = order_items_model.order_id")
                 .build(),
         )
-        .view(
-            ViewBuilder::new("Customer_view")
-                .statement(r#"select * from wrenai.public."Customers""#)
-                .build(),
-        )
-        .view(ViewBuilder::new("Revenue_orders").statement(r#"select "Order_id", sum("Price") from wrenai.public."Order_items" group by "Order_id""#).build())
-        .view(
-            ViewBuilder::new("Revenue_orders_alias")
-                .statement(r#"select "Order_id" as "Order_id", sum("Price") as "Totalprice" from wrenai.public."Order_items" group by "Order_id""#)
-                .build())
         .build()
 }
